feat: add activity 6

+++ app/Http/Controllers/ProductController.php
@@ -54,6 +54,7 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
+        Gate::authorize('view-product', $product);
         return view('products.show', compact('product'));
     }
 
@@ -62,6 +63,7 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
+        Gate::authorize('manage-product', $product);
         return view('products.edit', compact('product'));
     }
 
@@ -70,6 +72,7 @@ class ProductController extends Controller
      */
     public function update(Request $request,  Product $product)
     {
+        Gate::authorize('manage-product', $product);
         $validated = $request->validate([
             'name'  => ['required', 'string', 'max:255'],
             'price' => ['required', 'numeric', 'min:0'],
@@ -91,6 +94,7 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        Gate::authorize('manage-product', $product);
         $product->delete();
 
         return redirect()
